+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/client/api/BookKeeper.java
@@ -20,8 +20,7 @@
  */
 package org.apache.bookkeeper.client.api;
 
-import org.apache.bookkeeper.client.impl.BookKeeperBuilder;
-import org.apache.bookkeeper.client.BookKeeperBuilderImpl;
+import org.apache.bookkeeper.client.impl.BookKeeperBuilderImpl;
 import org.apache.bookkeeper.conf.ClientConfiguration;
 
 /**
